refactor:  date 변환 로직 dateUtils 파일로 분리
