+++ src/utils/dateUtils.js
@@ -1,0 +1,25 @@
+export const getTimeAgo = (dateString) => {
+  const now = new Date();
+  const date = new Date(dateString);
+
+  const diffInSeconds = Math.floor((now - date) / 1000);
+  const diffInMinutes = Math.floor(diffInSeconds / 60);
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  const diffInDays = Math.floor(diffInHours / 24);
+  const diffInMonths = Math.floor(diffInDays / 30);
+  const diffInYears = Math.floor(diffInDays / 365);
+
+  if (diffInSeconds < 60) {
+    return `방금 전`;
+  } else if (diffInMinutes < 60) {
+    return `${diffInMinutes}분 전`;
+  } else if (diffInHours < 24) {
+    return `${diffInHours}시간 전`;
+  } else if (diffInDays < 30) {
+    return `${diffInDays}일 전`;
+  } else if (diffInMonths < 12) {
+    return `${diffInMonths}달 전`;
+  } else {
+    return `${diffInYears}년 전`;
+  }
+};
